update
update class Genre, Member, MemberAuctionRelation

+++ moun/src/main/java/io/moun/api/domain/genre/Genre.java
@@ -15,7 +15,7 @@ import org.hibernate.annotations.Immutable;
 public class Genre {
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private Long id;
+        private int id;
 
         @Enumerated(EnumType.STRING)
         private GenreType genreType;
